requirements.txt
crewai
crewai-tools
gradio
python-dotenv
langchain-openai
ddgs
langchain-community
